+++ python/qualitylib/report.py
@@ -141,7 +141,8 @@ class QualityReport(object):
                            metric.SonarAnalysisAge)
     DEPENDENCY_METRIC_CLASSES = (metric.DependencyQuality,
                                  metric.SnapshotDependencies,
-                                 metric.OWASPDependencies)
+                                 metric.HighPriorityOWASPDependencyWarnings,
+                                 metric.NormalPriorityOWASPDependencyWarnings)
     PERFORMANCE_METRIC_CLASSES = (metric.ResponseTimes,
                                   metric.YmorResponseTimes)
     ENVIRONMENT_METRIC_CLASSES = (metric.FailingCIJobs,
@@ -327,8 +328,8 @@ class QualityReport(object):
         metrics.extend(self.__mandatory_subject_metrics(product, self.TEST_DESIGN_METRIC_CLASSES))
         metrics.extend(self.__optional_subject_metrics(product, self.CODE_METRIC_CLASSES +
                                                        self.PERFORMANCE_METRIC_CLASSES))
-        if metric.OWASPDependencies.should_be_measured(product):
-            metrics.append(metric.OWASPDependencies(product, project=self.__project))
+        metrics.extend(self.__mandatory_subject_metrics(product, (metric.HighPriorityOWASPDependencyWarnings,
+                                                                  metric.NormalPriorityOWASPDependencyWarnings)))
         if metric.SnapshotDependencies.can_be_measured(product, self.__project):
             metrics.append(metric.SnapshotDependencies(product, report=self, project=self.__project))
         metrics.extend(self.__art_metrics(product.art()))